# Improve performance of ForEachAlive